Created basic S3 class 'cvCovEst'.  Also created a list of arguments to be output with cvCovEst.  These can be called in plotting/summary functions later to ensure consistency of results.  For example, making sure data is appropriately scaled when generating heatmaps.
Created a function is.cvCovEst() to check for class "cvCovEst".  This can be used in the summary/plotting and any other future cvCovEst extensions.  Also can be used externally.

Changed the name of cvSummary() to be a generic method summary.cvCovEst().  Now users can simply call summary() on the output of cvCovEst, or they can pipe (cvCovEst %>% summary()).

Updated test-cvCovEst.R to reflect the additional output, 'args'.